nx_script

+++ nxwebserver.cpp
@@ -431,7 +431,7 @@ void page_names(AsyncWebServerRequest *request) {
 void page_script_post(AsyncWebServerRequest *request) {
    #ifdef ESP32
     w_preferences.begin("nx", false);
-    const AsyncWebParameter* j = request->getParam(0); // 1st parameter
+    const AsyncWebParameter* j = request->getParam((size_t)0); // 1st parameter
     w_preferences.putString("script", j->value());
     //w_preferences.putString("script", "nop static set");
     w_preferences.end();
